Initial work on improving the book.
